requirements.txt
fastapi
uvicorn
dbus-next
wget 
flask-socketio
sacn
flask
python-xlib
pillow
pygame
twitchApi 
dotenv
colorlog
simpleobsws
mysql-connector-python
num2words
aiofiles
requests
